feat: dockerize BorderScraper app
