+++ scraper/crontab.template
@@ -1,0 +1,7 @@
+SHELL=/bin/bash
+# This makes each cron job source the container env that was exported in start-cron.sh
+BASH_ENV=/app/container.env
+
+# Run scraper.py
+${SCRAPER_CAR_INTERVAL} root /usr/local/bin/python /app/scraper_car.py
+${SCRAPER_QUEUE_INTERVAL} root /usr/local/bin/python /app/scraper_queue.py
