+++ visualizer/visualizer_functions.py
@@ -1,15 +1,12 @@
 import os
 from dotenv import load_dotenv
-import requests
-from datetime import datetime
-import psycopg2
 import pandas as pd
 import seaborn as sns
 import matplotlib.pyplot as plt
+import psycopg2
 
 # Load environment variables from .env file
 load_dotenv()
-
 
 def get_database_connection():
     """Create a database connection using credentials from a .env file."""
@@ -20,42 +17,6 @@ def get_database_connection():
         host=os.getenv("DB_HOST"),
         port=os.getenv("DB_PORT")
     )
-
-
-def fetch_data_from_api(url):
-    """Fetch JSON data from the provided API."""
-    response = requests.get(url)
-    
-    if response.status_code == 200:
-        return response.json()
-    else:
-        raise Exception(f"Failed to fetch data: {response.status_code}, {response.text}")
-
-
-def insert_data_into_db(buffer_zone_id, query, data):
-    """Insert fetched data into the border_statistics table."""
-    
-    # Add buffer_zone_id to the data dictionary
-    data['buffer_zone_id'] = buffer_zone_id
-    
-    conn = None
-    try:
-        conn = get_database_connection()
-        with conn.cursor() as cursor:
-            cursor.execute(query, data)
-        conn.commit()
-    except Exception as e:
-        if conn:
-            conn.rollback()
-        print(f"Database error: {e}")
-    finally:
-        if conn:
-            conn.close()
-
-
-def convert_date(date_str):
-    return datetime.strptime(date_str, "%H:%M:%S %d.%m.%Y")
-
 
 def create_visual (intervals, cln_index, cln_columns, cln_values):
     
@@ -91,7 +52,7 @@ def create_visual (intervals, cln_index, cln_columns, cln_values):
         plt.grid()
         plt.tight_layout()
         
-        plt.savefig(f'{os.getenv("WORKING_DIR_PATH")}/www/queue_length_visual_{interval.replace(" ", "_")}.png')
+        plt.savefig(f'./www/queue_length_visual_{interval.replace(" ", "_")}.png')
         plt.close()
 
     conn.close()
