+++ www/index.html
@@ -38,15 +38,6 @@
 	}
     </style>
 </head>
-<!-- Google tag (gtag.js) -->
-<script async src="https://www.googletagmanager.com/gtag/js?id=G-7FLPQ7MFVW"></script>
-<script>
-  window.dataLayer = window.dataLayer || [];
-  function gtag(){dataLayer.push(arguments);}
-  gtag('js', new Date());
-
-  gtag('config', 'G-7FLPQ7MFVW');
-</script>
 <body>
     <h1>BY to LT border queue</h1>
     <h3>Here, you can find historical data on queues at border crossing points from Belarus to Lithuania.<br>
